+++ Tests/Utils/FunctionDeclarations/GetParametersDiffTest.php
@@ -11,7 +11,6 @@
 namespace PHPCSUtils\Tests\Utils\FunctionDeclarations;
 
 use PHPCSUtils\TestUtils\UtilityMethodTestCase;
-use PHPCSUtils\Tokens\Collections;
 use PHPCSUtils\Utils\FunctionDeclarations;
 
 /**
@@ -39,350 +38,6 @@ class GetParametersDiffTest extends UtilityMethodTestCase
         $this->expectPhpcsException('$stackPtr must be of type T_FUNCTION, T_CLOSURE or T_USE or an arrow function');
 
         FunctionDeclarations::getParameters(self::$phpcsFile, 10000);
-    }
-
-    /**
-     * Verify recognition of PHP8 union type declaration.
-     *
-     * @return void
-     */
-    public function testPHP8UnionTypesSimple()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 8, // Offset from the T_FUNCTION token.
-            'name'                => '$number',
-            'content'             => 'int|float $number',
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'int|float',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 6, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => 9,
-        ];
-        $expected[1] = [
-            'token'               => 17, // Offset from the T_FUNCTION token.
-            'name'                => '$obj',
-            'content'             => 'self|parent &...$obj',
-            'pass_by_reference'   => true,
-            'reference_token'     => 15, // Offset from the T_FUNCTION token.
-            'variable_length'     => true,
-            'variadic_token'      => 16,
-            'type_hint'           => 'self|parent',
-            'type_hint_token'     => 11, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 13, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 union type declaration with a bitwise or in the default value.
-     *
-     * @return void
-     */
-    public function testPHP8UnionTypesSimpleWithBitwiseOrInDefault()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 6, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'int|float $var = CONSTANT_A | CONSTANT_B',
-            'default'             => 'CONSTANT_A | CONSTANT_B',
-            'default_token'       => 10, // Offset from the T_FUNCTION token.
-            'default_equal_token' => 8, // Offset from the T_FUNCTION token.
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'int|float',
-            'type_hint_token'     => 2, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 4, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $arrowTokenTypes = Collections::arrowFunctionTokensBC();
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected, $arrowTokenTypes);
-    }
-
-    /**
-     * Verify recognition of PHP8 union type declaration with two classes.
-     *
-     * @return void
-     */
-    public function testPHP8UnionTypesTwoClasses()
-    {
-        $php8Names = parent::usesPhp8NameTokens();
-
-        $expected    = [];
-        $expected[0] = [
-            'token'               => ($php8Names === true) ? 8 : 11, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'MyClassA|\Package\MyClassB $var',
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'MyClassA|\Package\MyClassB',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => ($php8Names === true) ? 6 : 9, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 union type declaration with all base types.
-     *
-     * @return void
-     */
-    public function testPHP8UnionTypesAllBaseTypes()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 20, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'array|bool|callable|int|float|null|object|string $var',
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'array|bool|callable|int|float|null|object|string',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 18, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 union type declaration with all pseudo types..
-     *
-     * Note: "Resource" is not a type, but seen as a class name.
-     *
-     * @return void
-     */
-    public function testPHP8UnionTypesAllPseudoTypes()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 16, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'false|mixed|self|parent|iterable|Resource $var',
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'false|mixed|self|parent|iterable|Resource',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 14, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 union type declaration with (illegal) nullability.
-     *
-     * @return void
-     */
-    public function testPHP8UnionTypesNullable()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 8, // Offset from the T_FUNCTION token.
-            'name'                => '$number',
-            'content'             => '?int|float $number',
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => '?int|float',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 6, // Offset from the T_FUNCTION token.
-            'nullable_type'       => true,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 type declaration with (illegal) single type null.
-     *
-     * @return void
-     */
-    public function testPHP8PseudoTypeNull()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 6, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'null $var = null',
-            'default'             => 'null',
-            'default_token'       => 10, // Offset from the T_FUNCTION token.
-            'default_equal_token' => 8, // Offset from the T_FUNCTION token.
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'null',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 4, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 type declaration with (illegal) single type false.
-     *
-     * @return void
-     */
-    public function testPHP8PseudoTypeFalse()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 6, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'false $var = false',
-            'default'             => 'false',
-            'default_token'       => 10, // Offset from the T_FUNCTION token.
-            'default_equal_token' => 8, // Offset from the T_FUNCTION token.
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'false',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 4, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 type declaration with (illegal) type false combined with type bool.
-     *
-     * @return void
-     */
-    public function testPHP8PseudoTypeFalseAndBool()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 8, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'bool|false $var = false',
-            'default'             => 'false',
-            'default_token'       => 12, // Offset from the T_FUNCTION token.
-            'default_equal_token' => 10, // Offset from the T_FUNCTION token.
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'bool|false',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 6, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 type declaration with (illegal) type object combined with a class name.
-     *
-     * @return void
-     */
-    public function testPHP8ObjectAndClass()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 8, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'object|ClassName $var',
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'object|ClassName',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 6, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 type declaration with (illegal) type iterable combined with array/Traversable.
-     *
-     * @return void
-     */
-    public function testPHP8PseudoTypeIterableAndArray()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 10, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'iterable|array|Traversable $var',
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'iterable|array|Traversable',
-            'type_hint_token'     => 4, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 8, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
-    }
-
-    /**
-     * Verify recognition of PHP8 type declaration with (illegal) duplicate types.
-     *
-     * @return void
-     */
-    public function testPHP8DuplicateTypeInUnionWhitespaceAndComment()
-    {
-        $expected    = [];
-        $expected[0] = [
-            'token'               => 17, // Offset from the T_FUNCTION token.
-            'name'                => '$var',
-            'content'             => 'int | string /*comment*/ | INT $var',
-            'pass_by_reference'   => false,
-            'reference_token'     => false,
-            'variable_length'     => false,
-            'variadic_token'      => false,
-            'type_hint'           => 'int|string|INT',
-            'type_hint_token'     => 5, // Offset from the T_FUNCTION token.
-            'type_hint_end_token' => 15, // Offset from the T_FUNCTION token.
-            'nullable_type'       => false,
-            'comma_token'         => false,
-        ];
-
-        $this->getParametersTestHelper('/* ' . __FUNCTION__ . ' */', $expected);
     }
 
     /**
